pyproject.toml
[project]
name = "pangu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ai-models>=0.7.2",
    "ai-models-panguweather>=0.0.7",
    "cfgrib>=0.9.15.1",
    "gdown>=5.2.0",
    "netcdf4>=1.7.3",
    "numpy>=2.4.0",
    "onnxruntime-gpu>=1.23.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
    "torch>=2.9.1",
    "xarray>=2025.12.0",
    "cdsapi>=0.7.5",
    "gdex-api-client", # Expected to be installed in editable mode or added manually in environment
    "cdo>=1.6.1",
    "timm>=1.0.22",
    "ruff>=0.14.10",
    "radon>=6.0.1",
    "isort>=7.0.0",
]
